Source/Oyooni/Banknote Detection/requirements.txt
# pip install -r requirements.txt

numpy>=1.18.5
Pillow
PyYAML>=5.3.1
torch>=1.7.0
torchvision>=0.8.1
pandas
